+++ botbase/__init__.py
@@ -1,21 +1,13 @@
 from logging import NullHandler, getLogger
 
-from .botbase import BotBase, get_handler
+from .botbase import *
 from .checks import *
-from .emojis import Emojis
-from .exceptions import Blacklisted
-from .wraps import (
-    MyContext,
-    MyInter,
-    Wrap,
-    WrappedChannel,
-    WrappedMember,
-    WrappedThread,
-    WrappedUser,
-)
-from .models import CogBase
+from .emojis import *
+from .exceptions import *
+from .wraps import *
+from .models import *
 
-__version__ = "1.20.0"  # x-release-please-version
+__version__ = "1.20.6"  # x-release-please-version
 
 
 getLogger(__name__).addHandler(NullHandler())